🚚 move withAuth to middlewares
